Removed the Python SerializeContextTools smoke test
Registered a C++ SerializeContextTools test that validates the `--help`
argument.

Fixed the logic in ly_add_target to associate the `AZ_TEST_EXECUTABLE`
define with an executable target that depends on AzTests.
