+++ Code/Tools/SerializeContextTools/CMakeLists.txt
@@ -11,6 +11,21 @@ if (NOT PAL_TRAIT_BUILD_HOST_TOOLS)
 endif()
 
 ly_add_target(
+    NAME SerializeContextTools.Object OBJECT
+    NAMESPACE AZ
+    FILES_CMAKE
+        serializecontexttools_object_files.cmake
+    INCLUDE_DIRECTORIES
+        PUBLIC
+            .
+    BUILD_DEPENDENCIES
+        PRIVATE
+            AZ::AzCore
+            AZ::AzFramework
+            AZ::AzToolsFramework
+)
+
+ly_add_target(
     NAME SerializeContextTools EXECUTABLE
     NAMESPACE AZ
     FILES_CMAKE
@@ -23,4 +38,29 @@ ly_add_target(
             AZ::AzCore
             AZ::AzFramework
             AZ::AzToolsFramework
+            $<TARGET_OBJECTS:SerializeContextTools.Object>
 )
+
+if(PAL_TRAIT_BUILD_TESTS_SUPPORTED)
+    ly_add_target(
+        NAME SerializeContextTools.Tests EXECUTABLE
+        NAMESPACE AZ
+        FILES_CMAKE
+            serializecontexttools_tests_files.cmake
+        INCLUDE_DIRECTORIES
+            PRIVATE
+                .
+        BUILD_DEPENDENCIES
+            PRIVATE
+                AZ::AzCore
+                AZ::AzFramework
+                AZ::AzToolsFramework
+                AZ::AzTest
+                $<TARGET_OBJECTS:SerializeContextTools.Object>
+    )
+
+    ly_add_googletest(
+        NAME AZ::SerializeContextTools.Tests
+        TEST_COMMAND $<TARGET_FILE:AZ::SerializeContextTools.Tests> AzRunUnitTests
+    )
+endif()
