+++ Code/Tools/SerializeContextTools/Runner.h
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) Contributors to the Open 3D Engine Project.
+ * For complete copyright and license terms please see the LICENSE at the root of this distribution.
+ *
+ * SPDX-License-Identifier: Apache-2.0 OR MIT
+ *
+ */
+
+#pragma once
+
+namespace SerializeContextTools
+{
+    //! Outputs the SerializeContextTools help usage
+    void PrintHelp();
+
+    //! Runs the SerializeContext application hook with the command line parameters.
+    int LaunchSerializeContextTools(int argc, char** argv);
+}
