+++ Code/Tools/SerializeContextTools/serializecontexttools_object_files.cmake
@@ -1,0 +1,23 @@
+#
+# Copyright (c) Contributors to the Open 3D Engine Project.
+# For complete copyright and license terms please see the LICENSE at the root of this distribution.
+#
+# SPDX-License-Identifier: Apache-2.0 OR MIT
+#
+#
+
+set(FILES
+    Application.h
+    Application.cpp
+    Converter.h
+    Converter.cpp
+    Dumper.h
+    Dumper.cpp
+    SliceConverterEditorEntityContextComponent.h
+    SliceConverter.h
+    SliceConverter.cpp
+    Runner.h
+    Runner.cpp
+    Utilities.h
+    Utilities.cpp
+)
